% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{alr.transformation.inv}
\alias{alr.transformation.inv}
\title{alr.transformation.inv}
\usage{
alr.transformation.inv(alr, refSp)
}
\arguments{
\item{alr}{matrix of data (variables in columns, samples in rows)}

\item{refSp}{reference variable}
}
\value{
inverse transformed matrix
}
\description{
Function to inverse transform a matrix from addative log ratio (alr) to compostions
}
