% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{alr.transformation}
\alias{alr.transformation}
\title{alr.transformation}
\usage{
alr.transformation(m, refSp)
}
\arguments{
\item{m}{matrix of data (variables in columns, samples in rows)}

\item{refSp}{reference variable}
}
\value{
transformed matrix
}
\description{
Function to transform a matrix of compositions with addative log ratio (alr)
}
