% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{BLASSO}
\alias{BLASSO}
\title{BLASSO}
\usage{
BLASSO(X, P, Ys, Fs, ncpu, rmSp, vnames, seed = NULL)
}
\arguments{
\item{X}{matrix of data (variables in columns, measurements in rows)}

\item{P}{matrix of perturbation indicators}

\item{Ys}{matrix of response (variables in columns, measurements in rows)}

\item{Fs}{matrix of filter indicating whether the Y-X pair is used}

\item{ncpu}{number of CPU used}

\item{rmSp}{the index of variable removed for alr transformed data}

\item{vnames}{the name of species to format output}

\item{seed}{the seed}
}
\value{
a list with alpha, beta, and output in MDSINE's format
}
\description{
Function to estimate parameters using gradient matching bayesian lasso
}
