% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{clr.transformation}
\alias{clr.transformation}
\title{clr.transformation}
\usage{
clr.transformation(m)
}
\arguments{
\item{m}{matrix of data (variables in columns, samples in rows)}
}
\value{
transformed matrix
}
\description{
Function to transform a matrix of compositions with centered log ratio (clr)
}
