% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{cumSumScale}
\alias{cumSumScale}
\title{cumSumScale}
\usage{
cumSumScale(m, p = 0.5)
}
\arguments{
\item{m}{matrix of data (variables in columns, measurements in rows)}

\item{p}{quantile used for normalization (default: 0.5)}
}
\description{
Function to perform cumulative sum scaling (CSS)
}
