% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{EM}
\alias{EM}
\title{EM}
\usage{
EM(dat, meta, forceBreak = NULL, useSpline = TRUE, dev = 5,
  verbose = TRUE, refSp = NULL, min_iter = 30, max_iter = 100,
  epsilon = 0.01, ncpu = 10, seed = NULL, scaling = 1000)
}
\arguments{
\item{dat}{input data following MDSINE's OTU table (i.e. variables in rows and samples in columns)}

\item{meta}{metadata following MDSINE's metadata format}

\item{forceBreak}{force to break the trajectory to handle pulsed perturbation (or species invasion) (default: NULL)}

\item{useSpline}{use spline to smooth data and calculate gradients (default: TRUE)}

\item{dev}{deviation (dev * mad) from the median to be considered as outliers (default: 5)}

\item{verbose}{print more information (default: TRUE)}

\item{refSp}{reference OTU for addative log ratio transformation (default: selected by BEEM)}

\item{min_iter}{minimal number of iterations for the EM algorithm (default: 30)}

\item{max_iter}{maximal number of iterations for the EM algorithm (default: 100)}

\item{epsilon}{tolerance threshold to early stop the iterations (default: 0.01)}

\item{ncpu}{maximal number of CPUs used (default:10)}

\item{seed}{seed used in BLASSO (default:NULL)}

\item{scaling}{median total biomass to scale all biomass data (default:1000)}
}
\description{
Function to estimate biomass and parameters simultaneously
}
