% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{formatOutput}
\alias{formatOutput}
\title{formatOutput}
\usage{
formatOutput(alpha, beta, gamma = NULL, vnames)
}
\arguments{
\item{vnames}{variable names}
}
\value{
a data frame in MDSINE's output format
}
\description{
Function to convert parameter vector alpha and matrix beta to MDSINE's output format
}
