% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{inspectEM}
\alias{inspectEM}
\title{inspectEM}
\usage{
inspectEM(beem.obj, counts)
}
\arguments{
\item{beem.obj}{BEEM output list}

\item{counts}{counts data following MDSINE's OTU table}
}
\description{
Diagnose the EM process
}
