% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{param.infer}
\alias{param.infer}
\title{param.infer}
\usage{
param.infer(
  dat,
  metadata,
  biomass,
  forceBreak = NULL,
  dev = Inf,
  ncpu = 4,
  norder = 3,
  infer_flag = TRUE
)
}
\arguments{
\item{dat}{input data following MDSINE's OTU table}

\item{metadata}{metadata following MDSINE's metadata format}

\item{biomass}{biomass data following MDSINE's biomass data format}

\item{forceBreak}{force to break the trajectory to handle pulsed perturbation (or species invasion) (default: NULL)}

\item{dev}{deviation (dev * mad) from the median to be considered as outliers (default:Inf, no filtering)}

\item{ncpu}{maximal number of CPUs used (default:4)}

\item{norder}{order of spline basis (default: 3)}

\item{infer_flag}{run inference (default:TRUE)}
}
\description{
Independent function for inferring parameters with Bayesian Lasso
}
