% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{paramFromEM}
\alias{paramFromEM}
\title{paramFromEM}
\usage{
paramFromEM(
  beem.obj,
  counts,
  metadata,
  sparse = TRUE,
  forceBreak = NULL,
  ncpu = 4,
  enforceLogistic = FALSE,
  dev = 0.05,
  forceSkipWarning = FALSE
)
}
\arguments{
\item{beem.obj}{BEEM output list}

\item{counts}{counts data following MDSINE's OTU table}

\item{metadata}{metadata following MDSINE's metadata format}

\item{sparse}{use the sparse mode to estimate the parameters (default: TRUE)}

\item{forceBreak}{force to break the trajectory to handle pulsed perturbation (or species invasion) (default: NULL)}

\item{ncpu}{maximal number of CPUs used (default:4)}

\item{enforceLogistic}{re-estimate the self-interaction parameters (enforce to negative values)}

\item{dev}{tolerance in mse to select iterations for estimating parameters (default: 0.05)}

\item{forceSkipWarning}{Use with caution! Force to skip model fit check.}
}
\description{
Inferring parameters from BEEM results
}
