% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{postProcessBeta}
\alias{postProcessBeta}
\title{postProcessBeta}
\usage{
postProcessBeta(beta, thre_pos = 0.001, thre_neg = 0.001)
}
\arguments{
\item{beta}{interaction matrix estimated with BLASSO}

\item{thre_pos}{threshold used to remove small positive interactions (default 1e-3 * slef interaction)}

\item{thre_neg}{threshold used to remove small negative interactions (default 1e-3 * slef interaction)}
}
\description{
Function to filter small interactions. Small interaction is defined relative to the mean of the diagonal of the matrix
}
