% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{postProcessGamma}
\alias{postProcessGamma}
\title{postProcessGamma}
\usage{
postProcessGamma(alpha, gamma, thre = 0.01)
}
\arguments{
\item{alpha}{growth vector estimated with BLASSO}

\item{gamma}{perterbation effect matrix estimated with BLASSO}

\item{thre}{threshold used to remove small perturbation effect (default 1e-2 * self interactions)}
}
\description{
Function to filter small perturbation effect. Small interaction is defined relative to the mean of the diagonal of the matrix
}
