% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{preProcess}
\alias{preProcess}
\title{preProcess}
\usage{
preProcess(
  counts,
  metadata,
  rsp,
  dev = 100,
  scaling = 5000,
  smooth_data = TRUE,
  ncpu = 10,
  norder = 3,
  finite_diff = FALSE,
  forceBreak = NULL
)
}
\arguments{
\item{counts}{input data following MDSINE's OTU table (i.e. variables in rows and samples in columns)}

\item{metadata}{metadata following MDSINE's metadata format}

\item{rsp}{reference species used for alr transformation}

\item{dev}{deviation (dev * mad) from the median to be considered as outliers}

\item{scaling}{scale the normalized counts (default 5000)}

\item{smooth_data}{data will be smoothed before initial normalization (default: TRUE)}

\item{ncpu}{number of CPUs (default: 10)}

\item{norder}{order of spline basis (default: 3)}

\item{finite_diff}{use finite difference method to calculate gradients (default FALSE)}

\item{forceBreak}{force to break the trajectory to handle pulsed perturbation (or species invasion) (default: NULL)}
}
\description{
Function to preprocess input data. Input counts are normalized with CSS within each subject across all time points and then normalized across subjects with method described by David et. al.
}
