% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{smooth.deri.glkerns}
\alias{smooth.deri.glkerns}
\title{smooth.deri.glkerns}
\usage{
smooth.deri.glkerns(x, t, deriv = 0, t.out = t, smethod = "glkerns", ...)
}
\arguments{
\item{x}{vector of data}

\item{t}{time stamp of each point}

\item{deriv}{order of derivative to calculate}

\item{t.out}{time stamp of output (default: to be the same as input)}
}
\value{
a vector of fitted value after smoothing
}
\description{
Function to smooth data and calculate derivative with glkerns
}
