% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{SmoothGradient}
\alias{SmoothGradient}
\title{SmoothGradient}
\usage{
SmoothGradient(m, t, breaks, t.out = t, deriv = 1, ncpu = 10,
  method = "pspline", ...)
}
\arguments{
\item{m}{a matrix of data (variable in columns and time in rows)}

\item{t}{matrix or vector of time points for each data}

\item{breaks}{break points to re-smooth the data}

\item{t.out}{time stamps of output (default: to be the same as input)}

\item{deriv}{order of derivative to calculate}

\item{ncpu}{number of cpus (default: 10)}

\item{method}{smoothing method used to calculate gradients (default: pspline)}
}
\value{
a matrix of gradients calculated (deriv=1), a matrix of smoothed data (deriv=0)
}
\description{
Function to smooth data and calculate derivative with pspline
}
