% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{splitAt}
\alias{splitAt}
\title{splitAt}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{vetor of data}

\item{pos}{positions to split the data vector}
}
\value{
a list of split data
}
\description{
Function to split a vector into a list at specified positions. Adapted from http://stackoverflow.com/questions/16357962/r-split-numeric-vector-at-position
}
