% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{suggestRefs}
\alias{suggestRefs}
\title{suggestRefs}
\usage{
suggestRefs(dat, meta, scaling = 1)
}
\arguments{
\item{dat}{input data in relative abundances}

\item{meta}{metadata following MDSINE's metadata format}

\item{scaling}{library size to scale the data to proportions}
}
\description{
Function to select references for ALR
}
