% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{testStop}
\alias{testStop}
\title{testStop}
\usage{
testStop(x, epsilon = 0.002, win = 3)
}
\arguments{
\item{x}{a vector of mse trace}

\item{epsilon}{tolerance of mse}
}
\value{
a boolean value about the termination
}
\description{
Function to determine whether to stop the EM
}
