% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{tsoutliers}
\alias{tsoutliers}
\title{tsoutliers}
\usage{
tsoutliers(x, t, dev = 2)
}
\arguments{
\item{x}{vector of time series}

\item{t}{vector of time stamps}

\item{dev}{deviation (dev * mad) from the median to be considered as outliers}
}
\description{
Function to detect outliers in time serires
}
