% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{tsoutliers.w.brks}
\alias{tsoutliers.w.brks}
\title{tsoutliers.w.brks}
\usage{
tsoutliers.w.brks(x, t, breaks, dev = 2)
}
\arguments{
\item{x}{vector of time series}

\item{t}{vector of time stamps}

\item{breaks}{break points to re-smooth the data}

\item{dev}{deviation (dev * mad) from the median to be considered as outliers}
}
\description{
Function to detect outliers in time serires
}
