% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emFunctions.r
\name{biomassFromEM}
\alias{biomassFromEM}
\title{biomassFromEM}
\usage{
biomassFromEM(beem.obj, forceSkipWarning = FALSE)
}
\arguments{
\item{beem.obj}{BEEM output list}

\item{forceSkipWarning}{Use with caution! Force to skip model fit check.}
}
\description{
Inferring biomass from BEEM results
}
